requirements.txt
# Core dependencies
qdrant-client>=1.9.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
requests>=2.31.0
urllib3>=2.0.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0

# FastAPI dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# For development and testing(TODO)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0

# Logging and utilities
logging>=0.4.9.6
